add Kth checking -- separated files
